pyproject.toml
[tool.poetry]
name = "neptune"
version = "0.1.0"
description = "The BE app for the ChushaChurch project."
authors = ["Pavel Piasetski <pyasecky2012pavel@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.7.2"
attrs = "23.1.0"
beautifulsoup4 = "4.12.2"
build = "0.10.0"
CacheControl = "0.13.1"
certifi = "2023.7.22"
cffi = "1.15.1"
charset-normalizer = "3.2.0"
cleo = "2.0.1"
crashtest = "0.4.1"
cryptography = "41.0.3"
distlib = "0.3.7"
Django = "4.2.4"
djangorestframework = "3.14.0"
dulwich = "0.21.5"
filelock = "3.12.3"
idna = "3.4"
importlib-metadata = "6.8.0"
installer = "0.7.0"
jaraco-classes = "3.3.0"
jeepney = "0.8.0"
jsonschema = "4.17.3"
keyring = "24.2.0"
more-itertools = "10.1.0"
msgpack = "1.0.5"
packaging = "23.1"
pexpect = "4.8.0"
pkginfo = "1.9.6"
platformdirs = "3.10.0"
poetry = "1.6.1"
poetry-core = "1.7.0"
poetry-plugin-export = "1.5.0"
ptyprocess = "0.7.0"
pycparser = "2.21"
pyproject_hooks = "1.0.0"
pyrsistent = "0.19.3"
pytz = "2023.3"
rapidfuzz = "2.15.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
SecretStorage = "3.3.3"
shellingham = "1.5.3"
soupsieve = "2.4.1"
sqlparse = "0.4.4"
tomli = "2.0.1"
tomlkit = "0.12.1"
trove-classifiers = "2023.8.7"
typing_extensions = "4.7.1"
urllib3 = "2.0.4"
virtualenv = "20.24.4"
waitress = "2.1.2"
WebOb = "1.8.7"
WebTest = "3.0.0"
zipp = "3.16.2"
django-webtest = "^1.9.10"
pymysql = "^1.1.0"
structlog = "^24.1.0"


[[tool.poetry.source]]
name = "jaraco.classes"
url = "https://pypi.org/project/jaraco.classes/"
priority = "supplemental"


[tool.poetry.group.staging]
optional = true

[tool.poetry.group.staging.dependencies]
gunicorn = "^21.2.0"
dj-database-url = "^2.1.0"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
